Remove pagination from index and add archive list
I don't like how pagination is done by default in jekyll and I can no
longer use a plugin so don't do pages. The first three articles get
displayed in full text and the rest are links like in the archives. The
archives page still exists and is rendered the same as the second half
of the index.